+++ CHANGELOG.md
@@ -1,7 +1,33 @@
 # Change Log
-All notable changes to the "font-switcher" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## 4.0.0
 
-## [Unreleased]
-- Initial release
+- Add commands to change editor and terminal font sizes.
+- Add command to change terminal font family.
+- Add ability to add and remove fonts.
+- Remove `font-switcher.fontFamily`  and `font-switcher.enableLivePreview` settings.
+  - Live preview is always enabled and the extension uses the default `editor.fontFamily` setting.
+
+## 3.1.0
+
+- Fix icon resolution.
+
+## 3.0.0
+
+- Updated extension icon!
+
+## 2.0.0
+
+- Added live preview functionality. User has to explicitly enable this new setting.
+
+## 1.0.3
+
+- Added preview using up/down keys (similar to the "Select Theme" behavior)
+
+## 1.0.2
+
+- Fixed bug that would delete a font option if the menu was cancelled.
+
+## 1.0.0
+
+- Initial release of Font Switcher